Refactor project structure
Single package.json and config file
