+++ package.json
@@ -10,46 +10,20 @@
   "license": "Apache-2.0",
   "homepage": "https://github.com/okta/samples-nodejs-express-4",
   "scripts": {
-    "start": "node lib/index.js",
-    "test": "npm run test:docs && npm run test:e2e",
-    "mock-okta": "node tools/mock-okta",
-    "mock-okta:debug": "DEBUG=mock-okta npm run mock-okta",
-    "wrap": "node tools/wrap-dependencies",
-    "prestart": "build-frontend",
-    "test:e2e": "webdriver-manager update && protractor test/e2e/conf.js",
-    "test:mock-okta": "node test/app-server/lib/test-mock-okta",
-    "test:app-server": "mocha test/app-server/index --bail --timeout 30000",
-    "test:app-server-all": "mocha test/app-server/with-app-server-start --timeout 30000",
-    "test:docs": "mocha test/docs/specs/ --bail"
+    "postinstall": "node post-install.js",
+    "custom-login-server": "node custom-login/server.js",
+    "okta-hosted-login-server": "node okta-hosted-login/server.js",
+    "resource-server": "node resource-server/server.js",
+    "test": "echo \"Tests not yet implemented\""
   },
-  "devDependencies": {
-    "chalk": "1.1.3",
-    "cli-table2": "0.2.0",
-    "connect": "3.5.0",
-    "diff": "3.0.1",
-    "forever-monitor": "1.7.1",
-    "jasmine-reporters": "2.2.0",
-    "jws": "3.1.4",
-    "minimist": "1.2.0",
-    "protractor": "4.0.4",
-    "wait-on": "2.0.2",
-    "yakbak": "2.5.0"
-  },
+  "devDependencies": {},
   "dependencies": {
-    "@okta/oidc-middleware": "0.0.2",
-    "@okta/samples-js-angular-1": "git+https://github.com/okta/samples-js-angular-1.git",
-    "body-parser": "^1.18.2",
-    "chai": "3.5.0",
-    "chai-files": "1.4.0",
-    "chai-http": "3.0.0",
-    "consolidate": "0.14.1",
-    "cookie-parser": "1.4.3",
-    "express": "4.14.0",
-    "express-session": "^1.15.5",
-    "lodash.merge": "4.6.0",
-    "mocha": "3.1.2",
-    "mustache": "2.2.1",
-    "request": "2.78.0",
-    "oidc-token-hash": "^1.0.0"
+    "@okta/jwt-verifier": "0.0.11",
+    "@okta/oidc-middleware": "0.0.15",
+    "colors": "^1.1.2",
+    "cors": "^2.8.4",
+    "express": "^4.14.0",
+    "express-session": "^1.15.6",
+    "mustache-express": "^1.2.5"
   }
 }
